fix(kiloclaw): eliminate bootstrap orphan path
